+++ .eslintrc.json
@@ -9,6 +9,7 @@
     "es2021": true
   },
   "rules": {
-    "@typescript-eslint/no-explicit-any": "error"
+    "@typescript-eslint/no-explicit-any": "error",
+    "@next/next/no-img-element": "off"
   }
 }
